% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{rcPredict}
\alias{rcPredict}
\title{rcPredict}
\usage{
rcPredict(mdl, testingDf, classification = FALSE, mdlMethod)
}
\arguments{
\item{mdl}{input trained model}

\item{testingDf}{input testing data}

\item{classification}{boolean}

\item{mdlMethod}{string either lm, h2o or kerasdnn}
}
\value{
prediction is output
}
\description{
Predict from a rcissus model
}
\author{
Avants BB
}
