% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{rcTestingMatrix}
\alias{rcTestingMatrix}
\title{rcTestingMatrix}
\usage{
rcTestingMatrix(x, masks, rcb, patchRadius = 3, nsamples = 1000, seeds)
}
\arguments{
\item{x}{input filenames or image list defining one testing modality}

\item{masks}{input filenames or image list defining testing modality masks}

\item{rcb}{input image basis from \code{\link{rcBasis}}}

\item{patchRadius}{patch radius, integer value}

\item{nsamples}{number of samples per image}

\item{seeds}{random seeds for reproducibility across testing modalities}
}
\value{
testing matrix is output in list
}
\description{
Build a rcissus testing matrix
}
\author{
Avants BB
}
