% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{rcTrainingMatrix}
\alias{rcTrainingMatrix}
\title{rcTrainingMatrix}
\usage{
rcTrainingMatrix(y, x, masks, rcb, patchRadius = 3, nsamples = 1000, seeds)
}
\arguments{
\item{y}{input filenames or image list defining ground truth}

\item{x}{input filenames or image list defining one training modality}

\item{masks}{input filenames or image list defining training modality masks}

\item{rcb}{input image basis from \code{\link{rcBasis}}}

\item{patchRadius}{patch radius, integer value}

\item{nsamples}{number of samples per image}

\item{seeds}{random seeds for reproducibility across training modalities}
}
\value{
training matrix and vector (ground truth) is output in list
}
\description{
Build a rcissus training matrix
}
\seealso{
\code{\link[ANTsRCore]{ripmmarcPop}} \url{https://antsx.github.io/ANTsRCore/reference/ripmmarcPop.html}
}
\author{
Avants BB
}
