% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{rcTrainTranslation}
\alias{rcTrainTranslation}
\title{rcTrainTranslation}
\usage{
rcTrainTranslation(x, y, masks, nBases = 6, patchRadius = 3,
  meanCenter = FALSE, nsamples = 1000, seeds, hidden = c(200, 200),
  classification = FALSE, dataScaling = 1000, batchSize,
  mdlMethod = "h2o", epochs = 10)
}
\arguments{
\item{x}{input filenames or image list defining source modality}

\item{y}{input filenames or image list defining target modality}

\item{masks}{input filenames or image list defining source modality masks}

\item{nBases}{number of eigenvector bases}

\item{patchRadius}{patch radius, integer value}

\item{meanCenter}{boolean}

\item{nsamples}{number of samples per image}

\item{seeds}{random seeds for reproducibility across testing modalities}

\item{hidden}{Hidden layer sizes (e.g. \code{c(100,100)}). Defaults to \code{c(200,200)}.}

\item{classification}{boolean}

\item{dataScaling}{scaling by which to divide input data - results may be sensitive to this value}

\item{batchSize}{for training}

\item{mdlMethod}{string either lm, h2o or kerasdnn}

\item{epochs}{number of epochs (integer) over which to train}
}
\value{
list containing the image bases and the trained model and other relevant parameters.
}
\description{
Training for rcissus translation between image modalities
}
\author{
Avants BB
}
