% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{rcTranslate}
\alias{rcTranslate}
\title{rcTranslate}
\usage{
rcTranslate(x, rcmdl, mask, classification = FALSE)
}
\arguments{
\item{x}{input single image from source modality}

\item{rcmdl}{the trained model from \code{rcTrainTranslation}}

\item{mask}{optional input single image mask}

\item{classification}{boolean}
}
\value{
the translated image.
}
\description{
Rcissus translation between image modalities
}
\author{
Avants BB
}
