% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{readRcissus}
\alias{readRcissus}
\title{readRcissus}
\usage{
readRcissus(directoryname)
}
\arguments{
\item{directoryname}{to read}
}
\value{
rcissus list of objects
}
\description{
Read an rcissus basis set
}
\examples{
\dontrun{
pop = getANTsRData( "population" ) # list of example images
imgBases = rcBasis( pop )
tfn = tempfile()
writeRcissus( imgBases, 3, FALSE, directoryname = tfn )
imgB2 = readRcissus( tfn )
}
}
\author{
Avants BB
}
