% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{writeRcissus}
\alias{writeRcissus}
\title{writeRcissus}
\usage{
writeRcissus(basis, patchRadius, meanCenter, dataScaling = 1, deepNet = NA,
  directoryname)
}
\arguments{
\item{basis}{the basis image to write}

\item{patchRadius}{the radius associated with the basis}

\item{meanCenter}{boolean}

\item{dataScaling}{optional scaling value(s)}

\item{deepNet}{optional network}

\item{directoryname}{for output}
}
\value{
succeed or fail
}
\description{
Write an rcissus basis set
}
\examples{

\dontrun{
pop = getANTsRData( "population" ) # list of example images
imgBases = rcBasis( pop )
writeRcissus( imgBases, 3, FALSE, tempfile() )
}

}
\author{
Avants BB
}
