% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{rcBasis}
\alias{rcBasis}
\title{rcBasis}
\usage{
rcBasis(x, patchRadius = 3)
}
\arguments{
\item{x}{input filenames or image list}

\item{patchRadius}{patch radius, integer value}
}
\value{
basis is output
}
\description{
Learn a rcissus image basis
}
\examples{

pop = getANTsRData( "population" ) # list of example images
imgBases = rcBasis( pop )

}
\seealso{
\code{\link[ANTsRCore]{ripmmarcPop}} \url{https://antsx.github.io/ANTsRCore/reference/ripmmarcPop.html}
}
\author{
Avants BB
}
