% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{rcPredict}
\alias{rcPredict}
\title{rcPredict}
\usage{
rcPredict(mdl, testingDf)
}
\arguments{
\item{mdl}{input trained model}

\item{testingDf}{input testing data}
}
\value{
prediction is output
}
\description{
Predict from a rcissus model
}
\author{
Avants BB
}
