% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{rcTrain}
\alias{rcTrain}
\title{rcTrain}
\usage{
rcTrain(y, trainingDf, classification = FALSE, epochs = 200,
  max_mem = "100G")
}
\arguments{
\item{y}{outcome vector}

\item{trainingDf}{input training data}

\item{classification}{boolean}

\item{epochs}{number of epochs (integer) over which to train}

\item{max_mem}{sets maximum allowable memory for h2o deep learning}
}
\value{
model is output
}
\description{
Train a rcissus model with \code{lm} or deep learning from \code{h2o}
}
\seealso{
\code{\link[ANTsRCore]{ripmmarcPop}} \url{https://antsx.github.io/ANTsRCore/reference/ripmmarcPop.html}
}
\author{
Avants BB
}
