% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{rcTrain}
\alias{rcTrain}
\title{rcTrain}
\usage{
rcTrain(y, trainingDf)
}
\arguments{
\item{y}{outcome vector}

\item{trainingDf}{input training data}
}
\value{
model is output
}
\description{
Train a rcissus model
}
\seealso{
\code{\link[ANTsRCore]{ripmmarcPop}} \url{https://antsx.github.io/ANTsRCore/reference/ripmmarcPop.html}
}
\author{
Avants BB
}
